pyproject.toml
[tool.poetry]
name = "jobdam"
version = "0.1.1"
description = "Textual open chatting"
authors = ["Earthquakoo <cream5343@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
textual-dev = "^1.2.1"
requests = "^2.31.0"
websockets = "^12.0"

[tool.poetry.scripts]
jobdam="jobdam.ui.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
